+++ src/main/java/com/example/test_sql/service/MonHocService.java
@@ -2,10 +2,11 @@ package com.example.test_sql.service;
 
 import com.example.test_sql.dto.MonHocDTO;
 import com.example.test_sql.mapper.MonHocMapper;
-import com.example.test_sql.model.Khoa;
+import com.example.test_sql.model.Baithi;
 import com.example.test_sql.model.Monhoc;
 import com.example.test_sql.repository.MonHocRepository;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Component;
 
 
@@ -18,11 +19,12 @@ public class MonHocService {
     @Autowired
     MonHocRepository monHocRepository ;
 
+
     @Autowired
     MonHocMapper monHocMapper;
 
-    public List<MonHocDTO> getList(){
-        List<Monhoc> monhocList = monHocRepository.findAll();
+    public List<MonHocDTO> getList(Pageable pageable){
+        List<Monhoc> monhocList = monHocRepository.findAll(pageable).toList();
         List<MonHocDTO> monHocDTOList = new ArrayList<>();
         for (Monhoc monhoc : monhocList) {
             MonHocDTO monHocDTO = monHocMapper.monHocDTO(monhoc);
@@ -42,7 +44,7 @@ public class MonHocService {
         }
     }
 
-    public void post(MonHocDTO monHocDTO){
+    public MonHocDTO post(MonHocDTO monHocDTO){
         Optional<Monhoc> monhocPost = monHocRepository.findById(monHocDTO.getMaMonHoc());
         if(monhocPost.isPresent()){
             throw new RuntimeException("Đã tồn tại bản ghi " );
@@ -50,28 +52,31 @@ public class MonHocService {
         else{
             Monhoc monhoc = monHocMapper.monhoc(monHocDTO);
             monHocRepository.save(monhoc);
+            return monHocDTO;
         }
     }
 
-    public void put(MonHocDTO monHocDTO, String MaMonhoc){
-        Optional<Monhoc> monhocPut = monHocRepository.findById(MaMonhoc);
+    public MonHocDTO put(MonHocDTO monHocDTO){
+        Optional<Monhoc> monhocPut = monHocRepository.findById(monHocDTO.getMaMonHoc());
         if (monhocPut.isPresent()){
             Monhoc monhoc = monHocMapper.monhoc(monHocDTO);
             monhocPut.get().setTenMonHoc(monhoc.getTenMonHoc());
             monHocRepository.save(monhocPut.get());
+            return monHocDTO;
         }
         else {
-            throw new RuntimeException("Không tồn tại bản ghi " + MaMonhoc);
+            throw new RuntimeException("Không tồn tại bản ghi " + monHocDTO.getMaMonHoc());
         }
     }
 
     public void delete(String MaMonhoc){
         Optional<Monhoc> monhocDelete = monHocRepository.findById(MaMonhoc);
-        if (monhocDelete.isPresent()) {
+        List<Baithi> baithis = monhocDelete.get().getBaithis();
+        if (monhocDelete.isPresent() && baithis.size() == 0) {
             monHocRepository.deleteById(MaMonhoc);
         }
         else {
-            throw new RuntimeException("Không tồn tại bản ghi");
+            throw new RuntimeException("Không hợp lệ");
         }
     }
 }
